Add UrlReverseRouter with implementations to get URLs for controllers
The mock implementation provides simple predictable strings based on
provided parameters, so that they can be easily used in unit tests.

The real implementation uses the HttpMethodRoutes defined in the app
to find the canonical url for any given controller, and automatically
populate a url with the appropriate parameters / querystring args.