Log full exception for error reporting from DBALConnectionExceptionHandler
DBAL throws ConnectionException in a variety of situations, some of which
are more likely to be runtime application logic than transient network /
server issues.

For now, keep serving a maintenance page for all of them, but include
the exception in the log context so that it can trigger error reports /
monitoring while we review whether/how to make the handler more specific
in future.